Fix type

+++ src/main/java/org/yourcompany/yourproject/Entity/House.java
@@ -5,9 +5,9 @@ public class House {
     private String id;
     private Integer area;
     private String address;
-    private Integer streetInFrontOfHouse;
+    private Double streetInFrontOfHouse;
     private Double width;
-    private Integer height;
+    private Double height;
     private Integer floorNumber;
     private Integer bedroomNumber;
     private Integer bathroomNumber;
@@ -18,8 +18,9 @@ public class House {
     public House() {
     }
 
-    public House(Integer area, String address, Integer streetInFrontOfHouse, Double width,
-            Integer height, Integer floorNumber, Integer bedroomNumber,
+    public House(Integer area, String address, Double streetInFrontOfHouse, Double width,
+            Double height, Integer floorNumber,
+            Integer bedroomNumber,
             Integer bathroomNumber, String direction, Integer law, Double price) {
         this.area = area;
         this.address = address;
@@ -58,11 +59,11 @@ public class House {
         this.address = address;
     }
 
-    public Integer getStreetInFrontOfHouse() {
+    public Double getStreetInFrontOfHouse() {
         return streetInFrontOfHouse;
     }
 
-    public void setStreetInFrontOfHouse(Integer streetInFrontOfHouse) {
+    public void setStreetInFrontOfHouse(Double streetInFrontOfHouse) {
         this.streetInFrontOfHouse = streetInFrontOfHouse;
     }
 
@@ -74,11 +75,11 @@ public class House {
         this.width = width;
     }
 
-    public Integer getHeight() {
+    public Double getHeight() {
         return height;
     }
 
-    public void setHeight(Integer height) {
+    public void setHeight(Double height) {
         this.height = height;
     }
 
@@ -129,7 +130,7 @@ public class House {
     public void setPrice(Double price) {
         this.price = price;
     }
-    
+
     @Override
     public String toString() {
         return "House{" +
